% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCH.R
\name{createCH}
\alias{createCH}
\title{createCH}
\usage{
createCH(z, p.sum = 0.6, p.win = 0.4, sigma.y = 2, monthly = FALSE,
  cov = FALSE)
}
\arguments{
\item{z}{list containing the output from sim_data()}

\item{p.sum}{Summer detection probability}

\item{p.win}{Winter detection probability}

\item{sigma.y}{Observation error for count data}
}
\description{
Create capture history from z matrix
}

