% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_z.R
\name{sim_z}
\alias{sim_z}
\title{sim_z}
\usage{
sim_z(nInd = 150, omega = 65, nYears = 12, phi.sum = 0.97,
  phi.win = 0.98, mu.phi.aut = 0.9, lsigma.phi.aut, rel.diff, cor.spr)
}
\arguments{
\item{nInd}{Number of individuals added to population each year (if \code{staggered = FALSE}, nInd * nYears individuals are added in year 1)}

\item{nYears}{Total number of years in simulation (can be shortened in createCH)}

\item{phi.sum}{Monthly summer survival probability}

\item{phi.win}{Monthly winter survival probability}

\item{mu.phi.aut}{Mean monthly autumn survival probability}

\item{lsigma.phi.aut}{SD of variation in autumn survival}

\item{rel.diff}{Relative difference between spring and fall migration survival (i.e, mu.phi.spr = mu.phi.aut * rel.diff)}

\item{cor.spr}{Covariance between spring and fall migration}

\item{staggered}{Should nInd new individuals be added to simulation each year (if FALSE, all individuals added in year 1)}
}
\description{
Simulate survival data
}

