% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_z_cov.R
\name{sim_z_cov}
\alias{sim_z_cov}
\title{sim_z_cov}
\usage{
sim_z_cov(nInd = 50, nYears = 12, phi.sum = 0.97, phi.win = 0.98,
  mu.phi.aut = 0.9, r.var = 0.1, beta.aut, beta.spr, rel.diff,
  staggered = TRUE)
}
\arguments{
\item{nInd}{Number of individuals added to population each year (if \code{staggered = FALSE}, nInd * nYears individuals are added in year 1)}

\item{nYears}{Total number of years in simulation (can be shortened in createCH)}

\item{phi.sum}{Monthly summer survival probability}

\item{phi.win}{Monthly winter survival probability}

\item{mu.phi.aut}{Mean monthly autumn survival probability}

\item{rel.diff}{Relative difference between spring and fall migration survival (i.e, mu.phi.spr = mu.phi.aut * rel.diff)}

\item{staggered}{Should nInd new individuals be added to simulation each year (if FALSE, all individuals added in year 1)}

\item{lsigma.phi.aut}{SD of variation in autumn survival}

\item{cor.spr}{Covariance between spring and fall migration}
}
\description{
Simulate survival data
}

